# docs/openapi_admin.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "管理员端接口",
    "description": "管理员端的相关api。本地开发环境需要删除/api前缀再进行请求，服务器生产环境配置了nginx路由，直接请求即可。最终版。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/admin/auth/send-code": {
      "post": {
        "summary": "发送验证码",
        "deprecated": false,
        "description": "用于忘记密码或修改密码场景。",
        "tags": [
          "AdminAuth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "发送验证码参数",
                "type": "object",
                "properties": {
                  "email": {
                    "title": "接收邮箱",
                    "type": "string",
                    "format": "email",
                    "description": "目标邮箱地址"
                  }
                },
                "required": [
                  "email"
                ]
              },
              "example": {
                "email": "garhlz257@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "发送成功响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 200
                    },
                    "msg": {
                      "title": "提示信息",
                      "type": "string",
                      "example": "验证码已发送"
                    },
                    "data": {
                      "title": "业务数据",
                      "type": "object",
                      "properties": {},
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 200,
                  "msg": "验证码已发送",
                  "data": {}
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 40302,
                  "message": "无权限访问",
                  "data": null
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 40401,
                  "message": "该账号未注册",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
    "/api/admin/auth/logout": {
      "post": {
        "summary": "退出登录",
        "deprecated": false,
        "description": "登出系统。",
        "tags": [
          "AdminAuth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "退出成功响应",
                "type": "object",
                "properties": {
                  "code": {
                    "title": "状态码",
                    "type": "integer",
                    "example": 200
                  },
                  "msg": {
                    "title": "提示信息",
                    "type": "string",
                    "example": "已退出登录"
                  },
                  "data": {
                    "title": "业务数据",
                    "type": "object",
                    "properties": {},
                    "nullable": true
                  }
                },
                "required": [
                  "code",
                  "msg"
                ]
              },
              "example": {
                "code": 200,
                "msg": "已退出登录",
                "data": null
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "退出成功响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 200
                    },
                    "msg": {
                      "title": "提示信息",
                      "type": "string",
                      "example": "已退出登录"
                    },
                    "data": {
                      "title": "业务数据",
                      "type": "object",
                      "properties": {},
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 200,
                  "msg": "已退出登录",
                  "data": {}
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "退出失败响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 401
                    },
                    "msg": {
                      "title": "错误提示",
                      "type": "string",
                      "example": "未登录或Token过期"
                    },
                    "data": {
                      "title": "数据",
                      "type": "null"
                    }
                  }
                },
                "example": {
                  "code": 401,
                  "msg": "未登录或Token过期",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/test/register": {
      "post": {
        "summary": "测试：创建管理员用户",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "nickname",
                  "password",
                  "email"
                ]
              },
              "example": {
                "nickname": "elaine",
                "password": "123456",
                "email": "garhlz257@163.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/refresh": {
      "post": {
        "summary": "刷新token",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "refreshToken": "..."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isNewUser": {
                          "type": "null"
                        },
                        "hasPassword": {
                          "type": "null"
                        },
                        "userInfo": {
                          "type": "null"
                        },
                        "userSig": {
                          "type": "null"
                        },
                        "token": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "isNewUser",
                        "hasPassword",
                        "userInfo",
                        "userSig",
                        "token",
                        "refreshToken"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 200,
                  "message": "操作成功",
                  "data": {
                    "isNewUser": null,
                    "hasPassword": null,
                    "userInfo": null,
                    "userSig": null,
                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjE3LCJyb2xlIjoiVVNFUiIsIm5pY2tuYW1lIjoi5pig6K6w5rWL6K-V6LSm5Y-3Lea1i-ivleeJiCIsInR5cGUiOiJhY2Nlc3MiLCJleHAiOjE3NjY5MDk1NTR9.oHGva64ncj8Zw2v1YkonQzJNUJRokr5qnZ9JVYXYCX8",
                    "refreshToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjE3LCJ0eXBlIjoicmVmcmVzaCIsImV4cCI6MTc2ODg5Njc1NH0.vLRdH_bOujDVR41KLSASVPHJe7c8he0iIPRzLGAglQE"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "jwt": []
    }
  ]
}